Opa, bom dia.

+++ Mecanicas/Assets/Scripts/HealthBarScript.cs
@@ -8,12 +8,12 @@ public class HealthBarScript : MonoBehaviour
     public Image fill;
     public Gradient gradient;
     public Slider slider;
-    public void SetMaxHealth(int health){
+    public void SetMaxHealth(float health){
         slider.maxValue = health;
         slider.value = health;
         fill.color = gradient.Evaluate(1f);
     }
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
